fastapi/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.0
h11==0.14.0
idna==3.4
pathlib==1.0.1
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymssql==2.2.10
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.104.0
h11==0.14.0
idna==3.4
pathlib==1.0.1
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymssql==2.2.10
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
fastapi==0.104.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
pathlib==1.0.1
pinecone-client==2.2.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymssql==2.2.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
fastapi==0.104.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
pathlib==1.0.1
pinecone-client==2.2.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymssql==2.2.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dnspython==2.4.2
fastapi==0.104.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pathlib==1.0.1
pinecone-client==2.2.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymssql==2.2.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
yarl==1.9.2aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
cssselect==1.2.0
dnspython==1.16.0
fastapi==0.104.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gnews==0.3.6
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pathlib==1.0.1
Pillow==10.1.0
pinecone-client==2.2.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==3.12.3
pymssql==2.2.10
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.26.0
requests-file==1.5.1
safetensors==0.4.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.23.2
yarl==1.9.2
